dev docs, moving bigtable to own module
