+++ geomesa-bigtable/geomesa-bigtable-datastore/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>geomesa-hbase_2.11</artifactId>
+        <artifactId>geomesa-bigtable_2.11</artifactId>
         <groupId>org.locationtech.geomesa</groupId>
         <version>1.4.0-SNAPSHOT</version>
     </parent>
