+++ geomesa-bigtable/geomesa-bigtable-datastore/src/main/resources/META-INF/services/org.geotools.data.DataStoreFactorySpi
