+++ geomesa-bigtable/geomesa-bigtable-datastore/src/main/scala/org/locationtech/geomesa/bigtable/data/BigtableDataStore.scala
