+++ geomesa-bigtable/geomesa-bigtable-datastore/src/test/resources/hbase-site.xml
