+++ geomesa-bigtable/geomesa-bigtable-dist/src/main/assemblies/binary-release.xml
