+++ geomesa-bigtable/geomesa-bigtable-dist/src/main/assemblies/component.xml
