+++ geomesa-bigtable/geomesa-bigtable-gs-plugin/README.md
