+++ geomesa-bigtable/geomesa-bigtable-gs-plugin/pom.xml
