+++ geomesa-bigtable/geomesa-bigtable-spark-runtime/src/main/scala/org/locationtech/geomesa/ScalaDocs.scala
