+++ geomesa-bigtable/geomesa-bigtable-spark/src/main/java/org/locationtech/geomesa/bigtable/spark/BigtableTableRecordReader.java
