+++ geomesa-bigtable/geomesa-bigtable-spark/src/main/resources/META-INF/services/org.locationtech.geomesa.spark.SpatialRDDProvider
