+++ geomesa-bigtable/geomesa-bigtable-tools/bin/install-hbase.sh
