+++ geomesa-bigtable/geomesa-bigtable-tools/conf/hbase-site.xml
