+++ geomesa-bigtable/geomesa-bigtable-tools/conf/java-logging.properties
