+++ geomesa-bigtable/geomesa-bigtable-tools/pom.xml
@@ -3,7 +3,7 @@
 
     <parent>
         <groupId>org.locationtech.geomesa</groupId>
-        <artifactId>geomesa-hbase_2.11</artifactId>
+        <artifactId>geomesa-bigtable_2.11</artifactId>
         <version>1.4.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
