+++ geomesa-bigtable/geomesa-bigtable-tools/src/main/resources/org/locationtech/geomesa/bigtable/tools/ingest-libjars.list
