+++ geomesa-bigtable/geomesa-bigtable-tools/src/main/scala/org/locationtech/geomesa/bigtable/tools/BigtableIngestCommand.scala
